Update Detail confirm, sample data

+++ src/pages/Detail.jsx
@@ -40,6 +40,11 @@ const Detail = ({ expenses, setExpenses }) => {
       return alert("올바른 입력이 아니거나 입력이 없습니다.");
     }
 
+    /* 수정 진행 여부 */
+    const confirmed = confirm("수정을 완료하시겠습니까?");
+
+    if (!confirmed) return;
+
     const updatedItem = {
       id: filteredItem.id,
       date: date.current.value,
@@ -63,6 +68,10 @@ const Detail = ({ expenses, setExpenses }) => {
 
   const removeItem = (e) => {
     e.preventDefault();
+
+    const confirmed = confirm("삭제하시겠습니까?");
+    if (!confirmed) return;
+
     expenses.splice(filteredItemIndex, 1);
     localStorage.setItem("expenses", JSON.stringify(expenses));
     setExpenses(expenses);
